Update docstrings and create TODO to fix bug with input_shape
