+++ 3d_cnn_prometheus/learning/model/bayesian_vnet.py
@@ -60,7 +60,11 @@ def bayesian_vnet(input_shape=(280, 280, 280, 1), kernel_size=3,
                   activation="relu", padding="SAME", **kwargs):
     prior_std = kwargs.get("prior_std", 1)
     prior_fn = normal_prior(prior_std)
-    inputs = Input(input_shape)
+
+    # TODO Why Input transform input_shape from (16, 32,32,16,1) to (?, 16, 32,32,16,1) ?
+    # Input(input_shape[1:]) - temporary bugfix to enable training
+
+    inputs = Input(input_shape[1:])
     conv1, pool1 = down_stage(inputs, 16,
                               kernel_size=kernel_size,
                               activation=activation,
