+++ 3d_cnn_prometheus/preprocessing/preprocessing_links/normalize_images.py
@@ -13,7 +13,13 @@ class NormalizeImages(ChainLink):
             imgs_path = link_config['imgs_path']
             self._normalize_images(imgs_path)
 
-    def _normalize_images(self, imgs_path: str):
+    def _normalize_images(self, imgs_path: str) -> None:
+        """
+        For each image in directory apply normalizing operations
+        and save with the same file name.
+        :param imgs_path: path to directory with images
+        :return:
+        """
         for img_name in os.listdir(imgs_path):
             img_path = os.path.join(imgs_path, img_name)
             img = np.load(img_path)
@@ -21,10 +27,10 @@ class NormalizeImages(ChainLink):
             np.save(img_path, img_norm)
 
     @staticmethod
-    def standardize(data: np.array):
+    def standardize(data: np.array) -> np.array:
         """
         Transforms data to have mean 0 and std 1.
         :param data: non-standardized image to transform
-        :return standardized  image
+        :return standardized image
         """
         return (data - np.mean(data)) / np.std(data)
